server/utils/calendarUtils.test.ts
import { calendar, dateRow, formatDaysForDateRow, generateDays } from './calendarUtils'
import {} from '../testutils/factories/bedOccupancyRange'

describe('calendarUtils', () => {
  describe('calendar', () => {
    it('should return calendar', () => {
      expect(calendar()).toMatchStringIgnoringWhitespace(
        `<table cellspacing="0">
        <thead>${dateRow()}</thead>
      </table>`,
      )
    })
  })

  describe('dateRow', () => {
    it('should return dateRow', () => {
      expect(dateRow()).toMatchStringIgnoringWhitespace(`<th>Room/Bed</th>
      ${formatDaysForDateRow(new Date())}`)
    })
  })

  describe('generateDays', () => {
    it('should return a string of the date for the next 30 days', () => {
      const days = generateDays(new Date(2023, 5, 14))
      expect(days).toHaveLength(30)
      expect(days).toEqual([
        new Date(2023, 5, 14),
        new Date(2023, 5, 15),
        new Date(2023, 5, 16),
        new Date(2023, 5, 17),
        new Date(2023, 5, 18),
        new Date(2023, 5, 19),
        new Date(2023, 5, 20),
        new Date(2023, 5, 21),
        new Date(2023, 5, 22),
        new Date(2023, 5, 23),
        new Date(2023, 5, 24),
        new Date(2023, 5, 25),
        new Date(2023, 5, 26),
        new Date(2023, 5, 27),
        new Date(2023, 5, 28),
        new Date(2023, 5, 29),
        new Date(2023, 5, 30),
        new Date(2023, 6, 1),
        new Date(2023, 6, 2),
        new Date(2023, 6, 3),
        new Date(2023, 6, 4),
        new Date(2023, 6, 5),
        new Date(2023, 6, 6),
        new Date(2023, 6, 7),
        new Date(2023, 6, 8),
        new Date(2023, 6, 9),
        new Date(2023, 6, 10),
        new Date(2023, 6, 11),
        new Date(2023, 6, 12),
        new Date(2023, 6, 13),
      ])
    })
  })
})
